refactor: add var total_samples
here, we want to replace the expression that tells us how many
samples to make space for, with the variable total_samples, so
that when we create the array of storage for the samples, the
number of samples (in that var) is an integer. we do that using
the round()I function, which looks at the fractional part, and if
it's greater than or equal to 0.5, we will drop the fraction and
add one; if it's less, then we drop the fraction but -not- add 1.